Move mergeInterpreters() and inSameDirectory() to the right module.
