+++ src/client/pythonEnvironments/info/index.ts
@@ -172,6 +172,62 @@ export function updateInterpreter(interp: PythonInterpreter, other: PythonInterp
 }
 
 /**
+ * Combine env info for matching environments.
+ *
+ * Environments are matched by path and version.
+ *
+ * @param interpreters - the env infos to merge
+ * @param deps - functional dependencies
+ * @prop deps.areSameInterpreter - determine if 2 infos match the same env
+ * @prop deps.normalizeInterpreter - standardize the given env info
+ */
+export function mergeInterpreters(
+    interpreters: PythonInterpreter[],
+    deps: {
+        areSameInterpreter(i1: PythonInterpreter, i2: PythonInterpreter): boolean;
+        normalizeInterpreter(i: PythonInterpreter): void;
+        updateInterpreter(i: PythonInterpreter, o: PythonInterpreter): void;
+    }
+): PythonInterpreter[] {
+    return interpreters.reduce<PythonInterpreter[]>((accumulator, current) => {
+        const existingItem = accumulator.find((item) => deps.areSameInterpreter(current, item));
+        if (!existingItem) {
+            const copied: PythonInterpreter = { ...current };
+            deps.normalizeInterpreter(copied);
+            accumulator.push(copied);
+        } else {
+            deps.updateInterpreter(existingItem, current);
+        }
+        return accumulator;
+    }, []);
+}
+
+/**
+ * Determine if the given paths are in the same directory.
+ *
+ * @param path1 - one of the two paths to compare
+ * @param path2 - one of the two paths to compare
+ * @param deps - functional dependencies
+ * @prop deps.arePathsSame - determine if two filenames point to the same file
+ * @prop deps.getPathDirname - (like `path.dirname`)
+ */
+export function inSameDirectory(
+    path1: string | undefined,
+    path2: string | undefined,
+    deps: {
+        arePathsSame(p1: string, p2: string): boolean;
+        getPathDirname(p: string): string;
+    }
+): boolean {
+    if (!path1 || !path2) {
+        return false;
+    }
+    const dir1 = deps.getPathDirname(path1);
+    const dir2 = deps.getPathDirname(path2);
+    return deps.arePathsSame(dir1, dir2);
+}
+
+/**
  * Build a version-sorted list from the given one, with lowest first.
  */
 export function sortInterpreters(interpreters: PythonInterpreter[]): PythonInterpreter[] {
